requirements.txt
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
idna==3.10
pycparser==2.23
PyGithub==2.8.1
PyJWT==2.10.1
PyNaCl==1.6.0
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
